Question 1
This includes a descriptive summary for players.csv and sessions.csv.

An overview is provided of n_rows and n_cols. There are tables with descriptions of the variables in each dataset.
Next, descriptive statistics are described for each dataset. Potential issues/limitations based on these descriptive statistics are also discussed.

STILL NOT COMPLETED: information about how the data was collected.